requirements.txt
tensorflow
pandas 
dvc   # data version control
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML  # used to read the YAML file using python
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask   # web develpement
Flask-Cors
-e .   # used for setup mathod